DCV-1024 : Improve preview sql display panel (#2)
* DCV-1024 : Several Changes

Add copy icon.
Show loading icon when reusing a panel.
Clarify displayLocation setting.
Add keyboard shortcut for previewing SQL.

* DCV-1024 : Fix copying error message. Show short error message in title

* DCV-1024 : Several changes

Add option to disable the version check errors.
Add option to change the query limit.
Update the webview panel to not reload the html.
Save the title in the state for the webview panel.

* DCV-1024 : Allow selecting part of query for preview.

* Fix displayLocation not working

Co-authored-by: Robbie Ostermann <robertostermann5@gmail.com>