Fix bug for empty sqlFmtAdditionalParams
In the previous code, an empty string is passed as `sqlFmtAdditionalParamsSetting`
when `sqlFmtAdditionalParams` has no settings.

This causes sqlfmt to stop working in my environment.

According to https://docs.sqlfmt.com/getting-started/configuring-sqlfmt#configuration-reference,
an empty string is never passed as an option in sqlfmt.
So, empty strings should not be included in `sqlFmtAdditionalParamsSetting`.